refactor: change the folder location of the mocks
