+++ __mocks__/fixtures/course.ts
